requirements.txt
pyyaml==3.14
peewee==3.8.2
boto3==1.9.111
